refactor(s2s-cli): move templates
