+++ packages/s2s-cli/src/templates/__tests__/fixtures/copy-result.js
