+++ packages/s2s-cli/src/templates/index.js
